master: Covert example/param_time_fun.py -> tests/test_functions.py
